refactor(draw): dissolve InteractionManager into utils
