+++ src/plugins/draw/utils/interactions/createDeleteInteractions.ts
