+++ src/plugins/draw/utils/interactions/createEditInteractions/createCutInteractions/makeDraw.ts
