+++ src/plugins/draw/utils/interactions/createEditInteractions/createCutInteractions/style.ts
