+++ src/plugins/draw/utils/interactions/createEditInteractions/createLassoInteraction.ts
