+++ src/plugins/draw/utils/interactions/createEditInteractions/createMergeInteraction.ts
