+++ src/plugins/draw/utils/interactions/localSelector.ts
